// apis/token-api.ts
export const CreateTokenGrantEnum = {
    Admin: 'admin',
    CreateSpace: 'create_space',
    DeleteSpace: 'delete_space',
    SpaceAdmin: 'space_admin',
    CreateDirectory: 'create_directory',
    DeleteDirectory: 'delete_directory',
    DeleteDirectoryPermanent: 'delete_directory_permanent',
    MoveDirectory: 'move_directory',
    CopyDirectory: 'copy_directory',
    UploadFile: 'upload_file',
    UploadFileForce: 'upload_file_force',
    BeginUpload: 'begin_upload',
    BeginUploadForce: 'begin_upload_force',
    ConfirmUpload: 'confirm_upload',
    CreateSymlink: 'create_symlink',
    CreateSymlinkForce: 'create_symlink_force',
    DeleteFile: 'delete_file',
    DeleteFilePermanent: 'delete_file_permanent',
    MoveFile: 'move_file',
    MoveFileForce: 'move_file_force',
    CopyFile: 'copy_file',
    CopyFileForce: 'copy_file_force',
    DeleteRecycled: 'delete_recycled',
    RestoreRecycled: 'restore_recycled'
} as const;
export type CreateTokenGrantEnum = typeof CreateTokenGrantEnum[keyof typeof CreateTokenGrantEnum];